+++ bootstrap.php
@@ -1,4 +1,13 @@
 <?php
+/**
+ * Bootstraps the CMB2 process
+ *
+ * @category  WordPress_Plugin
+ * @package   CMB2
+ * @author    WebDevStudios
+ * @license   GPL-2.0+
+ * @link      http://webdevstudios.com
+ */
 
 /**
  * Fires when CMB2 is included/loaded