+++ init.php
@@ -1,5 +1,11 @@
 <?php
 /**
+ * @category     WordPress_Plugin
+ * @package      CMB2
+ * @author       WebDevStudios
+ * @license      GPL-2.0+
+ * @link         http://webdevstudios.com
+ *
  * Plugin Name:  CMB2
  * Plugin URI:   https://github.com/WebDevStudios/CMB2
  * Description:  CMB2 will create metaboxes and forms with custom fields that will blow your mind.
@@ -11,7 +17,7 @@
  *               Bill Erickson (@billerickson / billerickson.net)
  *               Andrew Norcross (@norcross / andrewnorcross.com)
  *
- * Version:      2.0.5
+ * Version:      2.0.6
  *
  * Text Domain:  cmb2
  * Domain Path:  languages
@@ -42,38 +48,66 @@
                   or things might explode!
 *************************************************************************/
 
-if ( ! class_exists( 'cmb2_bootstrap_205', false ) ) {
+if ( ! class_exists( 'CMB2_Bootstrap_206', false ) ) {
 
 	/**
-	 * Check for newest version of CMB
+	 * Handles checking for and loading the newest version of CMB2
+	 *
+	 * @since  2.0.0
+	 *
+	 * @category  WordPress_Plugin
+	 * @package   CMB2
+	 * @author    WebDevStudios
+	 * @license   GPL-2.0+
+	 * @link      http://webdevstudios.com
 	 */
-	class cmb2_bootstrap_205 {
+	class CMB2_Bootstrap_206 {
 
 		/**
 		 * Current version number
 		 * @var   string
 		 * @since 1.0.0
 		 */
-		const VERSION = '2.0.4';
+		const VERSION = '2.0.6';
 
 		/**
-		 * Current version hook priority
+		 * Current version hook priority.
 		 * Will decrement with each release
 		 *
 		 * @var   int
 		 * @since 2.0.0
 		 */
-		const PRIORITY = 9994;
+		const PRIORITY = 9993;
 
-		public static $single = null;
+		/**
+		 * Single instance of the CMB2_Bootstrap_206 object
+		 *
+		 * @var CMB2_Bootstrap_206
+		 */
+		public static $single_instance = null;
 
-		public static function go() {
-			if ( null === self::$single ) {
-				self::$single = new self();
+		/**
+		 * Creates/returns the single instance CMB2_Bootstrap_206 object
+		 *
+		 * @since  2.0.0
+		 * @return CMB2_Bootstrap_206 Single instance object
+		 */
+		public static function initiate() {
+			if ( null === self::$single_instance ) {
+				self::$single_instance = new self();
 			}
-			return self::$single;
+			return self::$single_instance;
 		}
 
+		/**
+		 * Starts the version checking process.
+		 * Creates CMB2_LOADED definition for early detection by other scripts
+		 *
+		 * Hooks CMB2 inclusion to the init hook on a high priority which decrements
+		 * (increasing the priority) with each version release.
+		 *
+		 * @since 2.0.0
+		 */
 		private function __construct() {
 			/**
 			 * A constant you can use to check if CMB2 is loaded
@@ -85,30 +119,39 @@ if ( ! class_exists( 'cmb2_bootstrap_205', false ) ) {
 			add_action( 'init', array( $this, 'include_cmb' ), self::PRIORITY );
 		}
 
+		/**
+		 * A final check if CMB2 exists before kicking off our CMB2 loading.
+		 * CMB2_VERSION and CMB2_DIR constants are set at this point.
+		 *
+		 * @since  2.0.0
+		 */
 		public function include_cmb() {
-			if ( ! class_exists( 'CMB2', false ) ) {
-
-				if ( ! defined( 'CMB2_VERSION' ) ) {
-					define( 'CMB2_VERSION', self::VERSION );
-				}
-
-				if ( ! defined( 'CMB2_DIR' ) ) {
-					define( 'CMB2_DIR', trailingslashit( dirname( __FILE__ ) ) );
-				}
-
-				$this->l10ni18n();
-
-				// Include helper functions
-				require_once 'includes/helper-functions.php';
-				// Now kick off the class autoloader
-				spl_autoload_register( 'cmb2_autoload_classes' );
-				// Kick the whole thing off
-				require_once 'bootstrap.php';
+			if ( class_exists( 'CMB2', false ) ) {
+				return;
 			}
+
+			if ( ! defined( 'CMB2_VERSION' ) ) {
+				define( 'CMB2_VERSION', self::VERSION );
+			}
+
+			if ( ! defined( 'CMB2_DIR' ) ) {
+				define( 'CMB2_DIR', trailingslashit( dirname( __FILE__ ) ) );
+			}
+
+			$this->l10ni18n();
+
+			// Include helper functions
+			require_once 'includes/helper-functions.php';
+
+			// Now kick off the class autoloader
+			spl_autoload_register( 'cmb2_autoload_classes' );
+
+			// Kick the whole thing off
+			require_once 'bootstrap.php';
 		}
 
 		/**
-		 * Load CMB text domain
+		 * Registers CMB2 text domain path
 		 * @since  2.0.0
 		 */
 		public function l10ni18n() {
@@ -128,6 +171,8 @@ if ( ! class_exists( 'cmb2_bootstrap_205', false ) ) {
 		}
 
 	}
-	cmb2_bootstrap_205::go();
 
-} // class exists check
+	// Make it so...
+	CMB2_Bootstrap_206::initiate();
+
+}