+++ tests/test-cmb-core.php
@@ -1,8 +1,16 @@
 <?php
+/**
+ * CMB2 core tests
+ *
+ * @package   Tests_CMB2
+ * @author    WebDevStudios
+ * @license   GPL-2.0+
+ * @link      http://webdevstudios.com
+ */
 
 require_once( 'cmb-tests-base.php' );
 
-class CMB2_Core_Test extends CMB2_Test {
+class Test_CMB2_Core extends Test_CMB2 {
 
 	/**
 	 * Set up the test fixture
@@ -62,6 +70,7 @@ class CMB2_Core_Test extends CMB2_Test {
 			'priority'     => 'high',
 			'show_names'   => true, // Show field names on the left
 			'show_on'      => array(), // Specific post IDs or page templates to display this metabox
+			'show_on_cb'   => null, // Callback to determine if metabox should display. Overrides 'show_on'
 			'cmb_styles'   => true, // Include cmb bundled stylesheet
 			'fields'       => array(),
 			'hookup'       => true,
@@ -106,7 +115,7 @@ class CMB2_Core_Test extends CMB2_Test {
 	}
 
 	/**
-	 * @expectedException CMB2_Test_Exception
+	 * @expectedException Test_CMB2_Exception
 	 */
 	public function test_set_metabox_after_offlimits() {
 		try {
@@ -114,7 +123,7 @@ class CMB2_Core_Test extends CMB2_Test {
 			$this->cmb->metabox['title'] = 'title';
 		} catch ( Exception $e ) {
 			if ( 'Exception' === get_class( $e ) ) {
-				throw new CMB2_Test_Exception( $e->getMessage(), $e->getCode() );
+				throw new Test_CMB2_Exception( $e->getMessage(), $e->getCode() );
 			}
 		}
 	}
@@ -202,7 +211,7 @@ class CMB2_Core_Test extends CMB2_Test {
 		<form class="cmb-form" method="post" id="' . $this->cmb_id . '" enctype="multipart/form-data" encoding="multipart/form-data">
 			<input type="hidden" name="object_id" value="' . $this->post_id . '">
 			' . wp_nonce_field( $this->cmb->nonce(), $this->cmb->nonce(), false, false ) . '
-			<!-- Begin CMB Fields -->
+			<!-- Begin CMB2 Fields -->
 			<div class="cmb2-wrap form-table">
 				<div id="cmb2-metabox-' . $this->cmb_id . '" class="cmb2-metabox cmb-field-list">
 					function test_before_row Description test_test
@@ -220,7 +229,7 @@ class CMB2_Core_Test extends CMB2_Test {
 					testing after row
 				</div>
 			</div>
-			<!-- End CMB Fields -->
+			<!-- End CMB2 Fields -->
 			<input type="submit" name="submit-cmb" value="Save" class="button-primary">
 		</form>
 		';
@@ -276,7 +285,7 @@ class CMB2_Core_Test extends CMB2_Test {
 	}
 
 	public function test_boxes_get() {
-		new CMB2_for_testing( $this->metabox_array2 );
+		new Test_CMB2_Object( $this->metabox_array2 );
 
 		// Retrieve the instance
 		$cmb = cmb2_get_metabox( 'test2' );
@@ -378,7 +387,7 @@ class CMB2_Core_Test extends CMB2_Test {
 
 	}
 
-	public function test_add_group_field() {
+	public function test_add_group_field( $do_assertions = null ) {
 
 		// Retrieve a CMB2 instance
 		$cmb = cmb2_get_metabox( 'test2' );
@@ -411,6 +420,86 @@ class CMB2_Core_Test extends CMB2_Test {
 		), 1 ); // Test that the position argument is working
 
 		$this->assertEquals( array( 'group_field', 'colorpicker' ), $sub_field_id );
+
+	}
+
+	public function test_group_field_param_callbacks() {
+
+		// Retrieve a CMB2 instance
+		$cmb = cmb2_get_metabox( 'test2' );
+
+		$field_id = $cmb->update_field_property( 'group_field', 'before_group', 'before_group output' );
+		$this->assertTrue( ! empty( $field_id ) );
+
+		$cmb->update_field_property( 'group_field', 'before_group_row', 'before_group_row output' );
+		$cmb->update_field_property( 'group_field', 'after_group_row', 'after_group_row output' );
+		$cmb->update_field_property( 'group_field', 'after_group', 'after_group output' );
+
+		$fields = $cmb->prop( 'fields' );
+		$field = new CMB2_Field( array(
+			'field_args'  => $fields['group_field'],
+			'object_type' => $cmb->object_type(),
+			'object_id'   => $cmb->object_id(),
+		) );
+
+		$expected_group_render = '
+		before_group output
+		<div class="cmb-row cmb-repeat-group-wrap">
+			<div class="cmb-td">
+				<div id="group_field_repeat" class="cmb-nested cmb-field-list cmb-repeatable-group" style="width:100%;">
+					<div class="cmb-row cmb-group-description">
+						<div class="cmb-th">
+							<h2 class="cmb-group-name">Group</h2>
+							<p class="cmb2-metabox-description">Group description</p>
+						</div>
+					</div>
+					before_group_row output
+					<div class="postbox cmb-row cmb-repeatable-grouping" data-iterator="0">
+						<button disabled="disabled" data-selector="group_field_repeat" class="dashicons-before dashicons-no-alt cmb-remove-group-row"></button>
+						<div class="cmbhandle" title="Click to toggle"><br></div>
+						<h3 class="cmb-group-title cmbhandle-title"><span></span></h3>
+						<div class="inside cmb-td cmb-nested cmb-field-list">
+							<div class="cmb-row cmb-type-colorpicker cmb2-id-group-field-0-colorpicker cmb-repeat-group-field">
+								<div class="cmb-th">
+									<label for="group_field_0_colorpicker">Colorpicker</label>
+								</div>
+								<div class="cmb-td">
+									<input type="text" class="cmb2-colorpicker cmb2-text-small" name="group_field[0][colorpicker]" id="group_field_0_colorpicker" value="#"/>
+								</div>
+							</div>
+							<div class="cmb-row cmb-type-group cmb2-id-group-field-0-first-field cmb-repeat-group-field">
+								<div class="cmb-th">
+									<label for="group_field_0_first_field">Field 1</label>
+								</div>
+								<div class="cmb-td"></div>
+							</div>
+							<div class="cmb-row cmb-remove-field-row">
+								<div class="cmb-remove-row">
+									<button disabled="disabled" data-selector="group_field_repeat" class="button cmb-remove-group-row alignright">Remove Group</button>
+								</div>
+							</div>
+						</div>
+					</div>
+					after_group_row output
+					<div class="cmb-row">
+						<div class="cmb-td">
+						<p class="cmb-add-row">
+						<button data-selector="group_field_repeat" data-grouptitle="" class="cmb-add-group-row button">Add Group</button>
+						</p>
+						</div>
+					</div>
+				</div>
+			</div>
+		</div>
+		after_group output
+		';
+
+		ob_start();
+		$cmb->render_group( $field->args );
+		// grab the data from the output buffer and add it to our $content variable
+		$rendered_group = ob_get_clean();
+
+		$this->assertHTMLstringsAreEqual( $expected_group_render, $rendered_group );
 
 	}
 
@@ -533,7 +622,7 @@ class CMB2_Core_Test extends CMB2_Test {
 /**
  * Simply allows access to the mb_defaults protected property (for testing)
  */
-class CMB2_for_testing extends CMB2 {
+class Test_CMB2_Object extends CMB2 {
 	public function get_metabox_defaults() {
 		return $this->mb_defaults;
 	}
@@ -545,7 +634,7 @@ class CMB2_for_testing extends CMB2 {
  *
  * @link http://stackoverflow.com/a/10744841
  */
-class CMB2_Test_Exception extends Exception {
+class Test_CMB2_Exception extends Exception {
 	public function __construct( $message = null, $code = 0, Exception $previous = null ) {
 		parent::__construct( $message, $code );
 	}